requirements.txt
flask
flask-sqlalchemy
flask-login
flask-cors
python-dotenv
requests
werkzeug